Move deprecated method to partial
