+++ src/NServiceBus.Core/EndpointConfiguration.cs
@@ -15,7 +15,7 @@ using Settings;
 /// <summary>
 /// Configuration used to create an endpoint instance.
 /// </summary>
-public class EndpointConfiguration : ExposeSettings
+public partial class EndpointConfiguration : ExposeSettings
 {
     /// <summary>
     /// Initializes the endpoint configuration builder.
@@ -77,21 +77,6 @@ public class EndpointConfiguration : ExposeSettings
     /// Access to the pipeline configuration.
     /// </summary>
     public PipelineSettings Pipeline { get; }
-
-    /// <summary>
-    /// Used to configure components in the container.
-    /// </summary>
-    [ObsoleteMetadata(
-        Message = "Registering services via NServiceBus is being deprecated. Instead, consider using a hosting framework such as Generic Host or Web Host to register services in the ServiceCollection.",
-        TreatAsErrorFromVersion = "11",
-        RemoveInVersion = "12")]
-    [Obsolete("Registering services via NServiceBus is being deprecated. Instead, consider using a hosting framework such as Generic Host or Web Host to register services in the ServiceCollection.. Will be treated as an error from version 11.0.0. Will be removed in version 12.0.0.", false)]
-    public void RegisterComponents(Action<IServiceCollection> registration)
-    {
-        ArgumentNullException.ThrowIfNull(registration);
-
-        Settings.Get<HostingComponent.Settings>().UserRegistrations.Add(registration);
-    }
 
     /// <summary>
     /// Configures the endpoint to be send-only.
