+++ src/NServiceBus.Core/obsoletes-v10.cs
@@ -621,5 +621,27 @@ namespace NServiceBus.Unicast.Messages
     }
 }
 
+namespace NServiceBus
+{
+    using System;
+    using Microsoft.Extensions.DependencyInjection;
+    using Particular.Obsoletes;
+
+    public partial class EndpointConfiguration
+    {
+        [ObsoleteMetadata(
+            Message = "Registering services via NServiceBus is being deprecated. Instead, consider using a hosting framework such as Generic Host or Web Host to register services in the ServiceCollection.",
+            TreatAsErrorFromVersion = "11",
+            RemoveInVersion = "12")]
+        [Obsolete("Registering services via NServiceBus is being deprecated. Instead, consider using a hosting framework such as Generic Host or Web Host to register services in the ServiceCollection.. Will be treated as an error from version 11.0.0. Will be removed in version 12.0.0.", false)]
+        public void RegisterComponents(Action<IServiceCollection> registration)
+        {
+            ArgumentNullException.ThrowIfNull(registration);
+
+            Settings.Get<HostingComponent.Settings>().UserRegistrations.Add(registration);
+        }
+    }
+}
+
 #pragma warning restore CS8632 // The annotation for nullable reference types should only be used in code within a '#nullable' annotations context.
 #pragma warning restore CS1591 // Missing XML comment for publicly visible type or member
